+++ changelog.md
@@ -1,5 +1,11 @@
 # Change log
 
+### 1.1.0 03.04.2021
+
+- Added: class BE\::IndentedOut
+- Fixed: AST\::Interface\::get_all_bases ()
+- Fixed: pseudo interfaces may inherit other pseudo interfaces
+
 ### 1.0.1 28.03.2021
 
 - Bugs fixed